set up channels for microbits, break up node.ts
